+++ packages/vtable-plugins/bundler.config.js
@@ -10,12 +10,20 @@ module.exports = {
   name: 'VTable.plugins',
   umdOutputFilename: 'vtable-plugins',
   rollupOptions: {
-    treeshake: true
+    treeshake: true,
+    plugins: [resolve({ browser: true })]
   },
-  globals: {},
+  globals: {
+    '@visactor/vtable': 'VTable',
+    '@visactor/vtable/es/vrender': 'VRender',
+    '@visactor/vtable-gantt': 'VTableGantt',
+    roughjs: 'Rough'
+  },
   external: [
     // ...其他外部依赖
     '@visactor/vtable',
-    '@visactor/vtable-gantt'
+    '@visactor/vtable/es/vrender',
+    '@visactor/vtable-gantt',
+    'roughjs'
   ]
 };