+++ packages/vtable/src/dataset/dataset.ts
@@ -540,7 +540,7 @@ export class Dataset {
         const record = this.records[i];
         if (!isNeedFilter || this.filterRecord(record)) {
           (this.filteredRecords as any[]).push(record);
-          this.processRecord(record);
+          record && this.processRecord(record);
         }
       }
     } else {
@@ -1089,7 +1089,7 @@ export class Dataset {
     this.filterRules = filterRules;
     this.filteredRecords = undefined;
     if (isResetTree) {
-      this.tree = {};
+      return this.setRecords(this.records);
     } else {
       for (const treeRowKey in this.tree) {
         for (const treeColKey in this.tree[treeRowKey]) {